Added new function to fetch and cache recent data.
Added a new function to fetch test data from the last <duration> seconds for a pair of sites. Using a duration rather than start/end times gives a nice label that can be used with memcache. Ideally the matrix can now ask for the last 10 minutes or so of data and now get a cached response most of the time.