Fixed no NNTSC data arriving bug
Fixed the bug where a request by ampy for NNTSC data would not complete. This was typically because the last NNTSC history message would be small and therefore was read into its internal buffer along with the previous message. The previous message told us there was more data, so we would call the receive_message function. This would block waiting for data to arrive, but there was no more data coming -- it was already sitting in our buffer!

Instead, we don't call receive_message unless a call to parse_message has failed, i.e. the current message is not entirely present in the buffer.