Changed API for NNTSC data to use stream id
API function calls for getting data for smokeping or muninbytes now take a
stream id as a parameter instead of a "source" and "dest".

Muninbytes streams are now identified by a switch, interface and direction
rather than using the switch and the NNTSC name string. Added new API functions
for getting the interfaces based on a switch and the directions based on an
interface. Left the old functions in for now, but they shouldn't be used.

get_* API functions no longer return the list of destinations in the event of
a None source and suchlike. This doesn't make sense anymore because we're using
streams and there are separate functions for getting this info anyway.