Make new NNTSC api work with threads
Fixes to ensure that the NNTSC API doesn't break or perform suboptimally if two
users make simultaneous requests using the same Connection -- this happens a
lot with amp-web, for instance.

Wrap our core data structures with locks -- this should fix the problem where
the same parser can be created twice.

Don't use a single network connection to NNTSC -- instead, create one whenever
we need to make a request and close it when we are done. This will prevent
problems where we have to wait for the response from a previous request before
sending the next one.

The lock stuff may be a bit overdone, but I'm erring on the side of caution.
It doesn't appear to have a significant performance penalty, but we can
revisit it if it turns out that it does.