Documented the new NNTSC API
Added nice python doc strings to much of the new NNTSC API. Hopefully should
make it a bit easier for people to follow, although the get_selection_option
stuff is a pain to explain.

Also fixed potential bug in _process_new_streams where we don't check if we
have a parser for the collection before trying to retrieve it from the map.