Muninbytes parser now returns a Mbps value
The format_data functions for all parsers now takes a streaminfo parameter --
this is to allow for proper calculation of "bps" metrics based on the
known measurement frequency. Turns out this wasn't necessary for Munin as the
value written to the RRD was actually "bytes per second" and not "bytes" as we
had originally thought.