Don't require source / user to get protocols
The protocol list is simply a list of all protocols that are associated with
at least one stream, rather than being a complicated dictionary of protocols
that are associated with streams for a given source / user. This is much
faster to initialise and search and allows us to implement the dropdown list
in amp-web as NOT being dependent on a source and user being selected.

Same dealio with the direction list.

Fixed bug where the lock wouldn't be released if we tried to query streaminfo
for an invalid stream id.