Added parsers for the remaining LPI metrics
get_selection_options has changed behaviour slightly -- if it is expecting a
"_requesting" parameter and you don't provide one, you're going to get an
empty list back instead of it trying to search for a stream that matches
all of the parameters passed in.

Also updated the documentation in lpibytes.py to better match the current
behaviour.