Allow data fetching functions to set the memcache key, with custom duration.
get_recent_data() and get_period_data() have different use cases and different caching requirements, and now get different caching behaviours. get_period_data() continues to use the start and end of the period as the key, but get_recent_data() now uses the duration and binsize, as the end is assumed to be "now". When this used explicit timestamps the cached entry was never hit, because now had moved on.

The duration of the caching has also been increased for get_recent_data(), depending on the duration. These are fairly arbitrary values but should be in the right ballpark at least.