LPI users require protocol and direction now
Previously, LPI users could be selected based solely on the source. Now we use
(source, protocol, direction) as the key to the users dictionary. This allows
us to implement dropdowns that only show the users that are using a particular
protocol easily.

This is much better, since users tend to only use a handful of protocols so
there won't be a stream for most user / protocol combinations. Limiting the
user selection based on protocol will make it much less likely that we will
choose a stream that doesn't exist.