ampy now queries for new streams
Previously, if a new stream was created in the NNTSC database, ampy would
never learn about it or allow users to query it. Now, before ampy looks at
its internal stream dictionary, it will query NNTSC for any new streams and
update the dictionary first.

This has the obvious downside of adding an RTT whenever someone changes
a dropdown option. It would be nicer to just have the streams arrive as
they are created but we can't really keep a persistent connection to NNTSC
at this point.