Minor performance improvements in ampy
Assume a lock is held whenever _load_collections is called. This means we don't
waste time releasing and then immediately reacquiring the collection lock.

When calling a get_data function, the stream is no longer required to be in
the local stream cache. This save us some locking and a lookup in a
(potentially) large streams dictionary.