Added get_related_streams to ampy
The purpose of this function is to find out what streams you should be using
to populate the tabs that will go alongside the main graph.

At the moment, the definition of a related stream is pretty simplistic. We
check all of the collections that have the same root module, e.g.
get_related_streams for amp-icmp will check all of the amp-* collections.

Checking the same root module generally works because the parameters that
define a stream tend to have the same name across different collections within
that module, e.g. all the amp streams will have a "source" and a "destination".
That's not true once you start looking at other collections -- the smokeping
destination is called "host", for example.

The other downside is that we need a copy of all the streams for each related
collection, so the first person to load a graph for a particular collection
is going to have to wait even longer while we query the database for the
streams from several collections. After that, the streams are cached locally
within ampy so it only affects the first user.