Added request_data function to each parser
This function replaces all the different calls to get_aggregate_columns,
get_group_columns and get_aggregate_functions that we used to have in
_get_data. It also means that parsers can use different request types, e.g.
percentiles vs aggregate, depending on the level of detail without cluttering
up _get_data.

Due to this change, the 'percentiles' detail level is no longer required for
amp-icmp. Instead, the collection can use 'full' and the parser will know
that it needs to make a percentile request.