Attempt to cache the streaminfo
Depending on how our web server handles our AJAX requests, we can end up in a
situation where we have multiple requests to NNTSC for the same stream
information. Now, we try to cache the stream information (which generally
doesn't change) in memcache if available, so we can just read it straight out
of there if our current process is requiring stream info.

Knowing what streams are available is a bit more challenging -- we still want
to cache this to deal with near-simultaneous queries but expire it reasonably
quickly so that we can show new streams on our graphs.

Ultimately, this should cut down on the amount of data being transferred
between NNTSC and ampy whenever a user loads a graph.