Use the cache in get_stream_info where possible
Stream info is now cached, so this is really easy to lookup rather than
doing the whole "what have I got? what is in the cache? ok, let's bring me in
line with what is in the cache..." thing that we normally do.

Unfortunately, we still have to deal with the case where memcache is unavailable
or the stream is not in the cache.