Split "streams" dictionary in nntsc.py based on collection
This approach is mainly to allow us to more easily detect when a stream is no
longer available in NNTSC -- i.e. it is no longer in the list of streams we are
getting from NNTSC -- and so should be removed from our local dictionaries.

This may also help with lookup speed when dealing with the smaller collections.

We now always enforce the STREAM_CHECK_FREQUENCY even if we are using
memcache. No point poking at caches and doing set operations if it is unlikely
that anything will have changed.