Related stream id lists now record the original stream
This is to enable callers to be able to match the related streams when the
list of streams passed in has more than one member. In particular, we will use
this to ensure that when changing tabs in amp-web, the properties of the line
drawn for the original stream are retained for the related one (such as line
colour or hidden status).

It does mean that the 'streamid' returned is no longer a list -- instead it is
a dictionary indexed by the original stream.