Made AmpTracerouteParser a subclass of AmpIcmpParser
In most cases, they behave exactly the same because their streams are defined
in exactly the same way: source, dest, address, packet size.

In updating the traceroute parser to support addresses in the stream definition,
I realised that I was just duplicating all the changes I had made to the icmp
parser earlier today. By subclassing the traceroute parser from the icmp one,
I've removed a lot of duplicate code.

Also, moved the source and destination management code out of the icmp parser
and into the base amp parser, as these should be common to all amp collections
in future.