Added an AmpDnsParser to ampy
Due to the large number of parameters that describe a AMP DNS stream,
I've limited the hierarchical dictionaries for stream lookups to
just the source, dest, query and address fields.

It is still possible to query for the other stream fields, but the
lookups may be a little slower due to them being stored as a single
dictionary entry, i.e. every stream matching the main 4 fields will
be iterated over to find those that satisfy the query.

Ultimately, we may want to move towards a more flexible stream lookup
system, especially once we support constructing views.