Added support for amp icmp/traceroute views that are v4 or v6 only
Also added functionality to create a group based on an event stream id.
The stream id is nominal and describes *one* of the streams that make
up the time series that the event was detected on. Using that stream, we
now have to determine the group that it belongs to and create groups and
views based on that.