Implemented get_graphtab_group for all remaining collections
This required replacing the STREAM split for AMP ICMP/traceroute with an
ADDRESS split. This means that stream ids are no longer part of a group
definition, as these were very hard to convert across collections for
the tab callback.

Removed get_graphtab_stream from all collections as it is no longer used.

Make sure we detect and try to handle the case where get_graphtab_group
cannot create an appropriate group. In particular, make sure the error
return is consistent across all collections as well as what the caller
is checking for.

Update the stream map for the tab collection when get_graphtab_group is
called. This is mainly so that amp-icmp and amp-traceroute can check
if the current packet size should be used or if it should fall back to
the default size.