Disable caching of streams (except for active streams)
This was breaking horribly if no-one looked at the graphs
for more than a day, i.e. enough time for the stream info
entries to expire from the cache. A new ampy process
would then not be able to find streaminfo for any of the
expired streams and would update the stream list to only
contain streams that hadn't yet expired (usually streams
that started after the expiry time).

The only way to fix this was to flush the cache, which was
less than ideal.

So far, not caching streams seems to have not had any adverse
effects on performance. Data requests by the website will
be a little slower to start with, if they are using a fresh
ampy process/thread, but it is not as though we're constantly
trying to request streams.