First commit of the new and improved ampy
Main focus is on tidying up the code and breaking it down into
sensible components, removing redundant and unused code.

Collection 'parsers' are being reworked to be more OO, with a
base collection class providing all the common code.

The multi-level dictionaries that were used by each collection
to manage the stream properties and the streams that match
them have been replaced with a generic streammanager class that
can be used by any collection.

View management has been simplified. The groups and views tables
should scale better due to adding a column for the collection,
which will hopefully make queries faster.

Carefully re-implemented the block handling code to try and
ensure that queried data is assigned to blocks correctly. At the
very least, it should be a little easier to track down block
bugs from now on.