Remove unneeded translate_group method from amp-dns
The parent class will now default to simply calling create_group_description
if not explicitly overridden. This is the correct option if there are
no collections that will translate to this one, as you should therefore
only be calling translate_group with properties from a group that already
matches your current collection.