Added new methods for querying amp meta-data
 * Added a method to fetch a list of available meshes
 * Added a method to fetch info about a particular site

Modified get_matrix_data to return a list of sources and a list of
destinations.

Ensure get_matrix_data has an up-to-date list of streams before
querying for streams.

Changed get_selection_options to take a list rather than a dictionary,
as this save us having to construct the dictionary within amp-web (requiring
more collection-specific code in amp-web). The dictionary will instead
be constructed automatically in ampy.