Fix data fetches that occasionally miss certain streams
The problem was caused by filter_active_streams being called with
the time range from the request, not including any extra blocks that
were added. If a stream is only active during the extra blocks, it
therefore won't be included in the query.

In the worst cases, this can lead to strange gaps in the graphs that
disappear as soon as you zoom in or out. Flushing the cache and reloading
would also fix the problem, because the problematic block will no longer
be in the "extra" block range so the missing stream will now be included
in the query.

All in all, a tricky little bug that is hard to replicate and test but
I think I've fixed it as best I can tell.