Fix bug where extra nulls would appear at the end of blocks
If the block wasn't nicely aligned with the measurement frequency,
we often ended up inserting extra nulls at the end of the block
whenever that block was the last in a query. These nulls get
cached and cause our graphs to have lots of little gaps even when
there are no missing measurements.

Mainly a problem with LPI graphs when zoomed in -- binsize is 30 vs
measurement frequency of 300. Choosing more sensible binsizes will
help avoid similar problems in the future.