Change add_group_to_view to support multiple groups
Starting to get some cases where we want to construct a view
using multiple groups and creating a new view for each group as we
add it is a little silly. All existing uses of add_group_to_view
have been updated to call the new function with a list
containing a single item.