Allow views to span multiple collections
This is mainly to allow us to combine multiple latency
measurements on a single graph, e.g. comparing TCP ping with ICMP
latency.

Implemented an amp-latency 'viewstyle' for this purpose, so all
requests for ICMP, DNS or TCP ping data should use 'amp-latency' as
the collection when operating on a view involving any of these
collections.