Fix bug preventing group removal from views
Specifically, for amp-latency views -- amp-web would make
an API call using amp-latency as the collection (because it has
no idea what real collection the group belonged to), which
would cause modify_view to fail when it couldn't find a
matching collection module.

Now we only look up the collection module on 'add' rather than
'del', as we don't need it anyway. Instead for 'del', we treat
the collection string as the viewstyle in which case 'amp-latency'
is what we want for all latency groups anyway.