Moved data fetching functions to collection.py
Moved a lot of internal data fetching and caching functions from
ampy.py to collection.py.

The main reason for this is so that amp-traceroute can override
the get_collection_history function to fetch monitor map data.
This data will be binned and cached differently to regular
historical time series, so it makes sense to be able to override
the existing behaviour.

This has the added benefit of removing a whole lot of "internal"
code from the file that describes the ampy API.