Reinstate special case for low frequency tests
Turns out we do need it, as some destinations are tested a lot
less frequently than others, e.g. root DNS servers vs Google DNS.

If we don't do this, we end up with large gaps between each data
point for the low freq series. Plus, address graphs are no longer
relevant in this branch, which was why we removed it in the first
place.