Create separate module for AS traceroute
We need to use a separate viewstyle for astraceroute to make
life easier for amp-web, so we have to create a separate
collection module as well.

Probably not a bad thing really -- inheritance means we don't
have too much code duplication and the separation will probably
make it easier to maintain AS path code in the future.