Fix bad aggregation on hop matrix cell links
FAMILY aggregation doesn't make sense when the cell links through
to an AS path graph.

For now, clicking on a hop count cell will take you to the IPv4
AS path for that source/dest pair. A better approach would be to
check if IPv4 is available and fall back to IPv6 if not...