Add support for amp-http test data
Added new function for collections: "convert_property". Unfortunately, some
stream properties (like URLs) need to be encoded to prevent them from
creating broken AJAX queries. This function allows collections to provide
code for converting the encoded properties back into their original format
so they can be matched against the properties stored in the database.

Fixed bug in error reporting for amp-dns -- a label that was meant to be
printed wasn't in scope at the time the error message is printed.