Minor tweaks to support amp-web matrix fixes
'detail' is now used for formatting fetched datapoints. This is so we don't
end up doing AS name lookups when we are fetching the most common AS path for
the matrix and tooltips.

Make sure every collection supports a 'basic' detail level.

Default aggregation for AS traceroute views has been changed from IPv4 to
FAMILY. This fixes the problem where a hop matrix tooltip would never fetch
IPv6 hop counts. Also fix legend labels for AS rainbow graphs to only show
'IPv4' rather than 'BOTH' in this case.