Update matrix group fetching to use the split parameter
Ensures the view returned for each matrix cell will redirect to a graph
with only the families or directions that are being currently shown on
the matrix itself.

This fixes the problem where clicking on a cell in an IPv4 only matrix
would take the user to a graph showing both IPv4 and IPv6 lines.