Separate aggregation method from the rest of the legend label
This will allow us to avoid displaying the aggregation method on the
graph tooltips, which would often lead to silly duplications when combined
with the shortlabel.

E.g., "foo to bar IPv4 IPv4"

Also added explicit text describing the protocol used for a latency line
so it is easier to distinguish ICMP lines from DNS or TCP lines.