Add API function for searching ASNs by substring
This is used for operating the dropdown selector on the event filtering panel
for choosing an AS to include / exclude / highlight.

We do this page-by-page to avoid overwhelming the browser when, for example,
we have to load all 77K+ ASNs.

Also fix bug where the ASN 'Private' would cause problems when trying to
lookup the matching AS name.