modify_view now takes a viewstyle parameter
With the addition of amp-udpstream, a collection no longer maps to a specific
viewstyle (amp-udpstream will be able to be amp-latency or amp-udpstream,
depending on which metric we're looking at). As a result, the caller will need
to specify which type of view they are looking to create when dealing with
amp-udpstream data.

Also enabled amp-udpstream in ampy.py