Only allow traceroute translations if previous graph has only one group
Previously, if there were multiple series on a graph it was not clear
what series would be shown by clicking on the AS path tab. The quick fix
is to only allow single-series graphs to include an AS path tab.

Long-term, it might be better to instead pop up a dialog asking the user
which series they want to see the AS path for, but that's a trickier
proposition.