Add code to support UDPstream latency and loss matrix
This has also required a slight change to the get_matrix_data() API
function, so that the caller can specify a viewstyle for the resulting
matrix views. This is because udpstream supports both amp-latency and
amp-udpstream viewstyles, so we need a way to ensure that the right view
style is used for the matrix.