Added ability to fetch a subset of the available stream options for a given collection + pre-selection.
This is mainly to support pagination on the amp-web modals. If
you have 100s of possibilties for a stream parameter, you probably
don't want to be loading all of those into a dropdown before
displaying it -- showing the first 30 and adding more as the user
scrolls is a lot better.