Fix bug where legend labels for amp-traceroute hop count graphs were wrong
We were incorrectly assuming that the graph was going to be a rainbow graph
so were enforcing the rule where we can only view one family at a time.

This doesn't happen any more -- you can't create a rainbow graph group that
has more than one family -- so removing this enforcement will mean we don't
incorrectly apply it to other graphs leveraging this collection (specifically
the hop count graphs, where we can have both families on the graph at once).