Cache matrix cell view ids
Every matrix refresh was hammering the views table trying to
fetch the appropriate view ID for every matrix cell, one cell at
a time.

Now we try to cache these view IDs (which are unlikely to change
in most instances), which should speed up matrix loads as well as
reduce some burden on the views database.